refactor(knowledge-store): rewire the indexer onto the facade
converge/project/rows build the store through the facade and read its
Outcome, so the driven consumer never reaches store internals.
